requirements.txt
flask
flask-cors
openai
python-dotenv
pydub
ffmpeg-python
openpyxl
